+++ owncloudApp/src/main/res/xml/settings_more.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ownCloud Android client application
+
+  Copyright (C) 2021 ownCloud GmbH.
+
+  This program is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License version 2,
+  as published by the Free Software Foundation.
+
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+-->
+<PreferenceScreen xmlns:app="http://schemas.android.com/apk/res-auto"
+    app:key="more_screen">
+
+    <Preference
+        app:iconSpaceReserved="false"
+        app:key="help"
+        app:title="@string/prefs_help" />
+    <Preference
+        app:iconSpaceReserved="false"
+        app:key="syncCalendarContacts"
+        app:summary="@string/prefs_sync_calendar_contacts_summary"
+        app:title="@string/prefs_sync_calendar_contacts" />
+    <Preference
+        app:iconSpaceReserved="false"
+        app:key="recommend"
+        app:title="@string/prefs_recommend" />
+    <Preference
+        app:iconSpaceReserved="false"
+        app:key="feedback"
+        app:title="@string/drawer_feedback" />
+    <Preference
+        app:iconSpaceReserved="false"
+        app:key="privacyPolicy"
+        app:title="@string/prefs_privacy_policy" />
+    <Preference
+        app:iconSpaceReserved="false"
+        app:key="imprint"
+        app:title="@string/prefs_imprint" />
+    <Preference
+        app:iconSpaceReserved="false"
+        app:key="about_app"
+        app:title="@string/about_title" />
+
+</PreferenceScreen>
