+++ owncloudApp/src/main/res/xml/settings_security.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ownCloud Android client application
+
+  Copyright (C) 2021 ownCloud GmbH.
+
+  This program is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License version 2,
+  as published by the Free Software Foundation.
+
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+-->
+<PreferenceScreen xmlns:app="http://schemas.android.com/apk/res-auto"
+    app:key="security_screen">
+
+    <CheckBoxPreference
+        app:iconSpaceReserved="false"
+        app:key="set_pincode"
+        app:title="@string/prefs_passcode" />
+    <CheckBoxPreference
+        app:iconSpaceReserved="false"
+        app:key="set_pattern"
+        app:title="@string/prefs_pattern" />
+    <CheckBoxPreference
+        app:iconSpaceReserved="false"
+        app:key="set_biometric"
+        app:title="@string/prefs_biometric" />
+    <CheckBoxPreference
+        app:iconSpaceReserved="false"
+        app:key="touches_with_other_visible_windows"
+        app:summary="@string/prefs_touches_with_other_visible_windows_summary"
+        app:title="@string/prefs_touches_with_other_visible_windows" />
+
+</PreferenceScreen>
